refactor(component): split resource and resource list for reusability
